\name{BIB.dat}
\alias{BIB.dat}
\docType{data}
\title{
 BIB.dat
}
\description{
   An example dataset of BIB design.
}
\usage{data(BIB.dat)}
\format{
  A data frame with 108 observations on the following 3 variables.
  \describe{
    \item{\code{Trt}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9}}
    \item{\code{Blk}}{a factor with levels \code{b1} \code{b10} \code{b11} \code{b12} \code{b2} \code{b3} \code{b4} \code{b5} \code{b6} \code{b7} \code{b8} \code{b9}}
    \item{\code{y}}{a numeric vector}
  }
}

\references{
Yuanzhen Lin, Xiaoyang Chen. R & ASReml-R Statistical Analysis Tutorial. China Forestry Publishing House. 2014
}
\examples{
data(BIB.dat)
## maybe str(BIB.dat) ; plot(BIB.dat) ...
}
\keyword{datasets}
