\name{G.data}
\alias{G.data}
\docType{data}
\title{
Gblup trait data
}
\description{
An example dataset for GBLUP, cotaining aim traits and other factors.
}
\usage{data("G.data")}
\format{
  A data frame with 6607 observations on the following 7 variables.
  \describe{
    \item{\code{ID}}{a factor with levels \code{26} \code{27} \code{28} \code{29} \code{30} \code{31} \code{38} \code{39} \code{40} \code{41} \code{42} \code{43} \code{44} \code{45} \code{46} \code{47} \code{48} \code{49} \code{50} \code{51} \code{52} \code{53} \code{54} \code{55} \code{56} \code{57} \code{58} \code{59} \code{60} \code{61} \code{62} \code{63} \code{64} \code{65} \code{66} \code{67} \code{68} \code{69} \code{70} \code{71} \code{72} \code{73} \code{74} \code{75} \code{76} \code{77} \code{78} \code{79} \code{80} \code{81} \code{82} \code{83} \code{84} \code{85} \code{86} \code{87} \code{88} \code{90} \code{91} \code{92} \code{93} \code{94} \code{95} \code{96} \code{97} \code{98} \code{99} \code{100} \code{101} \code{102} \code{103} \code{104} \code{105} \code{106} \code{107} \code{109} \code{110} \code{111} \code{112} \code{113} \code{114} \code{115} \code{116} \code{117} \code{118} \code{119} \code{120} \code{121} \code{122} \code{123} \code{124} \code{125} \code{126} \code{127} \code{128} \code{129} \code{130} \code{131} \code{132} \code{133} \code{134} \code{135} \code{137} \code{138} \code{139} \code{140} \code{141} \code{142} \code{143} \code{144} \code{145} \code{146} \code{147} \code{149} \code{150} \code{151} \code{152} \code{153} \code{154} \code{155} \code{156} \code{157} \code{158} \code{159} \code{160} \code{161} \code{162} \code{163} \code{164} \code{165} \code{166} \code{167} \code{168} \code{169} \code{170} \code{171} \code{172} \code{173} \code{174} \code{175} \code{176} \code{177} \code{178} \code{179} \code{180} \code{181} \code{182} \code{183} \code{184} \code{185} \code{186} \code{187} \code{188} \code{189} \code{190} \code{191} \code{192} \code{193} \code{194} \code{195} \code{196} \code{197} \code{198} \code{199} \code{200}}
    \item{\code{Female}}{a factor with levels \code{1} \code{4} \code{6} \code{8} \code{9} \code{10} \code{11} \code{12}}
    \item{\code{Male}}{a factor with levels \code{1} \code{3} \code{4} \code{9} \code{10} \code{12} \code{19} \code{20} \code{21}}
    \item{\code{Year}}{a factor with levels \code{2000} \code{2001} \code{2002}}
    \item{\code{Site}}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16}}
    \item{\code{t1}}{a numeric vector}
    \item{\code{t2}}{a numeric vector}
  }
}

\examples{
data(G.data)
## maybe str(G.data) ; plot(G.data) ...
}
\keyword{datasets}
