\name{G.marker}
\alias{G.marker}
\docType{data}
\title{
Gblup marker data
}
\description{
An example dataset for GBLUP, cotaining id and SNPs.
}
\usage{data("G.marker")}
\format{
  A data frame with 178 observations on the following 3462 variables.
  \describe{
    \item{\code{X}}{a numeric vector}
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
    \item{\code{X5}}{a numeric vector}
    \item{\code{X6}}{a numeric vector}
    \item{\code{X7}}{a numeric vector}
    \item{\code{X8}}{a numeric vector}
    \item{\code{X9}}{a numeric vector}
    \item{\code{X10}}{a numeric vector}


    \item{\code{X3450}}{a numeric vector}
    \item{\code{X3451}}{a numeric vector}
    \item{\code{X3452}}{a numeric vector}
    \item{\code{X3453}}{a numeric vector}
    \item{\code{X3454}}{a numeric vector}
    \item{\code{X3455}}{a numeric vector}
    \item{\code{X3456}}{a numeric vector}
    \item{\code{X3457}}{a numeric vector}
    \item{\code{X3458}}{a numeric vector}
    \item{\code{X3459}}{a numeric vector}
    \item{\code{X3460}}{a numeric vector}
    \item{\code{X3461}}{a numeric vector}
  }
}
\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016
}
\seealso{
Website for instant update: yzhlin-asreml.ys168.com
}

\examples{
data(G.marker)
## maybe str(G.marker) ; plot(G.marker) ...
}
\keyword{datasets}
