\name{group.plot}
\alias{group.plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot multicompare result
}
\description{
This function plots bars of the averages of treatments to compare. It uses the objects generated by a procedure of comparison like LSD, HSD, etc. It can also display the 'average' value over each bar in a bar chart.
}
\usage{
group.plot(object, x.lbls, y.lbls = NULL, y.zero = NULL, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object created by a test of comparison from agricolae
}
  \item{x.lbls}{
x axis label
}
  \item{y.lbls}{
y axis label
}
  \item{y.zero}{
y axis tick from zero (0, default), or we can set least y axis tick.
}
  \item{\dots}{
Parameters of the function barplot()
}
}
\author{
Yuanzhen Lin <yzhlinscau@163.com>
}
\references{
Yuanzhen Lin, Xiaoyang Chen. R & ASReml-R Statistical Analysis Tutorial. China Forestry Publishing House. 2014
}
\seealso{
Website for instant update: yzhlin-asreml.ys168.com
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
library(agricolae)
data(sweetpotato)
model<-aov(yield~virus,data=sweetpotato)
comparison<- LSD.test(model,"virus",alpha=0.01)
group.plot(comparison,"virus")

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ group.plot }
\keyword{ multicomparison }% __ONLY ONE__ keyword per line
