\name{lattice.pdata}
\alias{lattice.pdata}
\docType{data}
\title{
 lattice.pdata
}
\description{
   An example dataset for lattice plotting.
}
\usage{data(lattice.pdata)}
\format{
  A data frame with 256 observations on the following 4 variables.
  \describe{
    \item{\code{Fam}}{a numeric vector }
    \item{\code{Spacing}}{a numeric vector}
    \item{\code{ca}}{a numeric vector }
    \item{\code{rank}}{a numeric vector}
  }
}

\references{
Yuanzhen Lin. R & ASReml-R Statistics. China Forestry Publishing House. 2016
}
\examples{
data(lattice.pdata)
## maybe str(lattice.pdata) ; plot(lattice.pdata) ...
}
\keyword{datasets}
